Improve contrast of active time panel button
Because the contrast between the 'active' background color and regular
background color is so low, telling which button in the time panel is
active is not trivial (especially in the dark mode).

Swapping the foreground and background colors of the active button makes
it easy to tell it apart from the others at a glance, while preserving
the color theme.