Activity skipping + chains balance & fixes
This is broken. I don't know why, because sometimes it works, sometimes it doesn't, changing the skip loop and the new method in activity.service doesn't seem to fix anything.

And I forgot to do a separate branch for all this so here's a major change with a bunch of minor balance and fixes.

Ask away.